requirements.txt
textual>=0.45.0
PyYAML>=6.0